Tax year implicits moved to own class.
